Check data.json existence before deleting on BE

+++ app.py
@@ -90,7 +90,8 @@ def remove_previous_data():
     with os.scandir(os.path.join('static', 'imgs')) as it:
         for entry in it:
             os.remove(entry.path)
-    os.remove('data.json')
+    if not os.path.exists('data.json'):
+        os.remove('data.json')
 
 
 def parse_time_str(time_str):
